requirements.txt
# numpy
# pandas
# matplotlib
# nltk
# scikit-learn
# ipykernel
# requests

python-dotenv
streamlit

# Local packages
-e .